refactor(zone): rename observability module to metrics
