+++ crates/tempo-zone/src/l1.rs
@@ -73,7 +73,7 @@ pub struct L1Subscriber {
     /// TIP-403 metrics (cache sizes, events applied).
     tip403_metrics: crate::l1_state::tip403::Tip403Metrics,
     /// L1 subscriber metrics for connection health, backfill, and event ingestion.
-    subscriber_metrics: crate::observability::L1SubscriberMetrics,
+    subscriber_metrics: crate::metrics::L1SubscriberMetrics,
 }
 
 impl L1Subscriber {
