+++ crates/tempo-zone/src/lib.rs
@@ -15,8 +15,8 @@ pub mod evm;
 mod executor;
 pub mod l1;
 pub mod l1_state;
+mod metrics;
 mod node;
-mod observability;
 pub mod payload;
 pub mod precompiles;
 pub mod rpc;
